Add installer action and customization by defines.
If the define "define_generate_documentation_only_for_first_build" is
set to "true" (not case sensitive), the documentation is only generated
if "define_build_is_the_first_one_for_artifact" is also set to true.
This introduces a way to generate the documentation only once for
artifacts with multiple builds.
The default is to generate the documentation for each build.

If the define "define_test_configuration" is set, it is used as the
path to the test configuration.
The default is "tests.xml".